requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pypika-tortoise==0.5.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
tortoise-orm>=0.17.8
pydantic>=1.8.2
email-validator>=1.1.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
aiosqlite>=0.17.0
bcrypt>=3.2.0